pyproject.toml
[tool.poetry]
name = "micro-ddd"
version = "0.1.0"
description = ""
authors = ["mohsen padidar <mxpadidar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
fastapi = "^0.110.1"
pydantic-settings = "^2.2.1"
uvicorn = "^0.29.0"
psycopg = "^3.1.18"
sqlalchemy = "^2.0.29"
httpx = "^0.27.0"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
alembic = "^1.13.1"


[tool.poetry.group.auth_service.dependencies]
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }


[tool.poetry.group.storage_service.dependencies]
minio = "^7.2.5"
python-magic = "^0.4.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
mypy_path = '$MYPY_CONFIG_FILE_DIR/'
